feat(core,legacy): add support for Ethereum 64-bit chain_id
* Changes from original PR

* Now that we are rejecting chain_ids of 0, we need to have the tests set the chain_ids to at least 1.

* Ran 'make gen' and uploaded changed files.

* Ran make style_check and fixed reported errors

* Added changelog files

* Reverted changes concerning chain_id 0 being rejected.

* Adds tests for MAX_CHAIN_ID and MAX_CHAIN_ID+1.  Also reverts MAX_CHAIN_ID to the previous value.

* Added missing whitespace around arithmetic operator.

+++ crypto/address.h
@@ -34,7 +34,7 @@ void address_write_prefix_bytes(uint32_t address_type, uint8_t *out);
 bool address_check_prefix(const uint8_t *addr, uint32_t address_type);
 #if USE_ETHEREUM
 void ethereum_address_checksum(const uint8_t *addr, char *address, bool rskip60,
-                               uint32_t chain_id);
+                               uint64_t chain_id);
 #endif
 
 #endif
